requirements.txt
rumps>=0.4.0
keyring>=23.0.1
cryptography>=41.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pynput>=1.7.6
